Full Crud with Authentication and Authorization

+++ Backend/controllers/goalController.js
@@ -1,6 +1,7 @@
 const asyncHandler = require("express-async-handler")
 const mongoose=require("mongoose")
-const Goal=require("../model/goalModel")
+const Goal = require("../model/goalModel")
+const User=require("../model/userModel")
 ///@desc Get goals
 // @route GET /api/goals
 //@access
@@ -9,9 +10,9 @@ const Goal=require("../model/goalModel")
 
 
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find()
+  const goals = await Goal.find({user:req.user._id})
   
-    res.status(200).json(goals);
+  res.status(200).json(goals);
 })
 
 
@@ -29,7 +30,8 @@ const setGoal =asyncHandler( async (req, res) => {
   }
 
   const goal = await Goal.create({
-    text:req.body.text
+    text: req.body.text,
+    user : req.user._id
   })
   res.status(200).json(goal);
 });
@@ -49,6 +51,15 @@ const updateGoal = asyncHandler(async (req, res) => {
     throw new Error("Goal not found")
   }
 
+  const user = await User.findById(req.user._id)
+  if (!user) {
+    res.status(401)
+    throw new Error("User not found")
+  }
+  if (goal.user.toString() != user._id) {
+    res.status(401)
+    throw new Error("User not authorized")
+  }
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, { new: true })
   
   res.status(200).json(updatedGoal);
@@ -68,6 +79,15 @@ const deleteGoal = asyncHandler(async (req, res) => {
     throw new Error("Goal not found")
   }
 
+  const user = await User.findById(req.user._id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+  if (goal.user.toString() != user._id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
   const deletedGoal=await Goal.findByIdAndDelete(req.params.id)
   res.status(200).json(deletedGoal);
 });
